+++ packages/client/src/widgets/leaderboard/index.ts
@@ -1,0 +1,3 @@
+import LeaderBoardWidget from './ui/LeaderBoardWidget'
+
+export default LeaderBoardWidget
