+++ packages/client/src/widgets/leaderboard/ui/LeaderBoardWidget/index.ts
@@ -1,0 +1,3 @@
+import LeaderBoardWidget from './LeaderBoardWidget'
+
+export default LeaderBoardWidget
